+++ aidge-20260428/alibabacloud_aidge20260428/models/_document_translate_response_body.py
@@ -18,13 +18,11 @@ class DocumentTranslateResponseBody(DaraModel):
         self.code = code
         # The asynchronous task information.
         self.data = data
-        # The error message. This parameter is not returned if the call is successful.
+        # The error message. Not returned for successful calls.
         self.message = message
         # Id of the request
         self.request_id = request_id
-        # Indicates whether the call was successful. Valid values:
-        # - true: The call was successful.
-        # - false: The call failed.
+        # Indicates whether the call is successful. Valid values: true: The call is successful. false: The call failed.
         self.success = success
 
     def validate(self):
@@ -78,7 +76,7 @@ class DocumentTranslateResponseBodyData(DaraModel):
         self,
         task_id: str = None,
     ):
-        # The unique identifier of the asynchronous task. Use this ID to query the task status and result.
+        # The unique identifier of the asynchronous task, used to query the task status and result.
         self.task_id = task_id
 
     def validate(self):